# cmd/server/docs/swagger.yaml
basePath: /
definitions:
  admin.ServerInfo:
    properties:
      version:
        example: 0.0.1
        type: string
    type: object
  iptables.initRequest:
    properties:
      chain_name:
        example: KNOCK_FW
        type: string
      exempt_ports:
        example:
        - "7999"
        - "7999"
        items:
          type: string
        type: array
      parent_chain:
        description: string or []string
        example:
        - INPUT
        - DOCKER-USER
        items:
          type: string
        type: array
    type: object
  iptables.ipRequest:
    properties:
      ip:
        example: 192.168.1.100
        type: string
    type: object
  models.AuthConfig:
    properties:
      auth_cache_expire:
        description: Cache expiration in seconds (default 60)
        example: 60
        type: integer
      auth_port:
        description: Local Auth Service Port
        example: 3000
        type: integer
      auth_url:
        description: Relative Verify URL (default /api/auth/verify)
        example: /api/auth/verify
        type: string
      login_url:
        description: Relative Login URL (default /login)
        example: /login
        type: string
      logout_url:
        description: Relative Logout URL (default /api/auth/logout)
        example: /api/auth/logout
        type: string
    type: object
  models.Rule:
    properties:
      path:
        description: Path prefix to match (e.g., "/api")
        example: /api
        type: string
      rewrite_html:
        description: If true, rewrites absolute paths in HTML response to include
          Path prefix.
        example: true
        type: boolean
      strip_path:
        description: If true, strips the Path prefix from the request before forwarding.
        example: true
        type: boolean
      target:
        description: Target URL (e.g., "http://localhost:7996")
        example: http://localhost:8080
        type: string
      use_auth:
        description: If true, invokes global authentication check before proxying.
        example: false
        type: boolean
      use_root_mode:
        description: If true, sets cookie and redirects matched path to /.
        example: false
        type: boolean
    type: object
  models.SSLInfo:
    properties:
      enabled:
        type: boolean
    type: object
  models.SSLRequest:
    properties:
      cert:
        example: |-
          -----BEGIN CERTIFICATE-----
          ...
        type: string
      key:
        example: |-
          -----BEGIN RSA PRIVATE KEY-----
          ...
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
host: 127.0.0.1:7996
info:
  contact: {}
  description: API for managing proxy rules and iptables.
  license:
    name: MIT
    url: https://opensource.org/license/MIT
  title: Go-Reauth-Proxy
  version: "1.0"
paths:
  /api/auth:
    get:
      description: Get the configured global authentication URLs and port
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuthConfig'
              type: object
      summary: Get global auth config
      tags:
      - config
    post:
      consumes:
      - application/json
      description: Set the global authentication configurations (port, auth_url, login_url)
      parameters:
      - description: Auth configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/models.AuthConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Set global auth config
      tags:
      - config
  /api/config/default-route:
    get:
      description: Get the configured default route when root route is requested
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get default route
      tags:
      - config
    post:
      consumes:
      - application/json
      description: Set the configured default route when root route is requested
      parameters:
      - description: 'Route configuration, example: {\'
        in: body
        name: rule
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Set default route
      tags:
      - config
  /api/info:
    get:
      description: Get version and other server info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.ServerInfo'
              type: object
      summary: Get server info
      tags:
      - info
  /api/iptables/allow:
    post:
      consumes:
      - application/json
      description: Add an ALLOW rule for a specific IP
      parameters:
      - description: IP to allow
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/iptables.ipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Allow IP
      tags:
      - iptables
  /api/iptables/allow-all:
    post:
      description: Remove the catch-all DROP rule
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Allow all traffic
      tags:
      - iptables
  /api/iptables/block:
    post:
      consumes:
      - application/json
      description: Add a DROP rule for a specific IP
      parameters:
      - description: IP to block
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/iptables.ipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Block IP
      tags:
      - iptables
  /api/iptables/block-all:
    post:
      description: Add a catch-all DROP rule at the end of the chain
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Block all traffic
      tags:
      - iptables
  /api/iptables/clean:
    post:
      description: Remove the custom iptables chain
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Clean iptables
      tags:
      - iptables
  /api/iptables/flush:
    post:
      description: Flush all rules in the custom chain
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Flush iptables rules
      tags:
      - iptables
  /api/iptables/init:
    post:
      consumes:
      - application/json
      description: Initialize the custom iptables chain
      parameters:
      - description: Initialization options
        in: body
        name: request
        schema:
          $ref: '#/definitions/iptables.initRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Initialize iptables
      tags:
      - iptables
  /api/iptables/list:
    get:
      description: List all rules in the custom chain
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: List iptables rules
      tags:
      - iptables
  /api/iptables/remove:
    post:
      consumes:
      - application/json
      description: Remove an ALLOW/BLOCK rule for a specific IP
      parameters:
      - description: IP to remove
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/iptables.ipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Remove IP rule
      tags:
      - iptables
  /api/rules:
    delete:
      description: Remove all proxy rules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Flush all rules
      tags:
      - rules
    get:
      description: Get all configured proxy rules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Rule'
                  type: array
              type: object
      summary: Get all rules
      tags:
      - rules
    post:
      consumes:
      - application/json
      description: Set proxy rules (overrides existing rules)
      parameters:
      - description: List of rules to set
        in: body
        name: rules
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Rule'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Rule'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Set rules
      tags:
      - rules
  /api/ssl:
    delete:
      description: Clear the configured SSL certificate and disable HTTPS on the proxy
        port
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Clear SSL certificate
      tags:
      - ssl
    get:
      description: Check if dynamic SSL is currently enabled and configured on the
        proxy port
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SSLInfo'
              type: object
      summary: Get SSL status
      tags:
      - ssl
    post:
      consumes:
      - application/json
      description: Upload a PEM encoded certificate and private key to enable HTTPS
        on the proxy port
      parameters:
      - description: SSL Certificate and Key
        in: body
        name: ssl
        required: true
        schema:
          $ref: '#/definitions/models.SSLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Set SSL certificate
      tags:
      - ssl
swagger: "2.0"